+++ src/main/java/edu/online/messenger/model/entity/User.java
@@ -1,15 +1,14 @@
 package edu.online.messenger.model.entity;
 
 import edu.online.messenger.constant.RoleName;
+import edu.online.messenger.model.entity.parent.BaseEntity;
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
+import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -23,11 +22,8 @@ import java.util.Set;
 @Setter
 @Entity
 @NoArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+@Table(name = "users")
+public class User extends BaseEntity {
 
     @Column(length = 50, nullable = false, unique = true)
     private String login;